% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cnv.R
\name{read_cnv}
\alias{read_cnv}
\alias{read_cnv_colmeta}
\alias{read_cnv_header}
\title{Read Seabird .cnv files}
\usage{
read_cnv(file, skip = 0, n_max = Inf, header = read_cnv_header(file))

read_cnv_colmeta(file, header = read_cnv_header(file))

read_cnv_header(file, n_header = 500)
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{skip}{Number of non-header rows to skip.}

\item{n_max}{Maximum number of rows to read.}

\item{header}{A previously read value obtained from \code{\link[=read_cnv_header]{read_cnv_header()}}.}

\item{n_header}{The starting guess for number of header lines.}
}
\description{
Read Seabird .cnv files
}
\examples{
file <- system.file("extdata/faroe/faroe.cnv", package = "sbe")
read_cnv(file)
read_cnv_colmeta(file)
head(read_cnv_header(file))

# header information is kept on read with the 'header' attribute
cnv <- read_cnv(file)
head(attr(cnv, "header"))

}
