% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sbe-cnv.R
\name{read_sbe_cnv}
\alias{read_sbe_cnv}
\alias{read_sbe_cnv_colmeta}
\alias{read_sbe_header}
\title{Read Seabird .cnv files}
\usage{
read_sbe_cnv(
  file,
  skip = 0,
  n_max = Inf,
  header = read_sbe_header(file),
  col_names = read_sbe_cnv_colmeta(file, header)$name,
  widths = rep(11L, length(col_names))
)

read_sbe_cnv_colmeta(file, header = read_sbe_header(file))

read_sbe_header(file, n_header = 500)
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{skip}{Number of non-header rows to skip.}

\item{n_max}{Maximum number of rows to read.}

\item{header}{A previously read value obtained from \code{\link[=read_sbe_header]{read_sbe_header()}}.}

\item{col_names}{The column names. Defaults to the \code{name} column of
\code{\link[=read_sbe_cnv_colmeta]{read_sbe_cnv_colmeta()}}.}

\item{widths}{A vector the same length as \code{col_names} denoting the
widths of each field in characters. Defaults to 11.}

\item{n_header}{The starting guess for number of header lines.}
}
\description{
The .cnv file format is the primary file format used for input and output
to SBE Data Processing. The format is essentially a fixed-width format using
11 characters for each column.
}
\examples{
file <- system.file("extdata/faroe/faroe.cnv", package = "sbe")
read_sbe_cnv(file)
read_sbe_cnv_colmeta(file)
head(read_sbe_header(file))

# header information is kept on read with the 'header' attribute
cnv <- read_sbe_cnv(file)
head(attr(cnv, "header"))

}
