% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sbe-hex.R
\name{read_sbe_hex}
\alias{read_sbe_hex}
\title{Read Seabird .hex files}
\usage{
read_sbe_hex(file, skip = 0, n_max = Inf, header = read_sbe_header(file))
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{skip}{Number of non-header rows to skip.}

\item{n_max}{Maximum number of rows to read.}

\item{header}{A previously read value obtained from \code{\link[=read_sbe_header]{read_sbe_header()}}.}
}
\description{
Read Seabird .hex files
}
