% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe-package.R
\docType{package}
\name{sbe-package}
\alias{sbe}
\alias{sbe-package}
\title{sbe: Process and Convert Files Created by 'Seabird' Products}
\description{
Provides facilities to process and convert 'Seabird'
  .hex, .dat, .xmlcon, .con files. An interface to the SBEBatch.exe
  batch processor is also provided.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
