% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbe_variables_tbl}
\alias{sbe_variables_tbl}
\title{Seabird Variable Information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 544 rows and 5 columns.
}
\usage{
sbe_variables_tbl
}
\description{
A table of variables commonly found in Seabird .cnv files. Useful in
conjunction with \code{\link[=read_sbe_cnv_colmeta]{read_sbe_cnv_colmeta()}}.
}
\keyword{datasets}
