% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbe-batch.R
\name{sbe_run_batch}
\alias{sbe_run_batch}
\alias{sbe_batch_path}
\alias{has_sbe_batch}
\alias{sbe_batch_configure}
\title{Run SBEBatch.exe}
\usage{
sbe_run_batch(batch_file, ..., quiet = FALSE, timeout = Inf)

sbe_batch_path(x)

has_sbe_batch()

sbe_batch_configure(reconfigure = FALSE, drive = NULL)
}
\arguments{
\item{batch_file}{A batchfile description.}

\item{...}{Additional arguments passed to the batch file processor.
Typically this will be a working directory. Use \code{\link[=sbe_batch_path]{sbe_batch_path()}}
to convert file paths that are valid in R to file paths that are valid
for SBEBatch.}

\item{quiet}{Use \code{TRUE} to suppress non-error output.}

\item{timeout}{See \code{\link[processx:run]{processx::run()}}.}
}
\description{
Run SBEBatch.exe
}
